fix(PostsEditor): correctly set images in the content and correctly add banner when creating posts
The regex for finding the images wasn't working as the local urls generated didn't match the pattern set in the regex.
As a solution, a Map was created to store the localUrl and its corresponding file name.

For the initial upload of posts, the banners were expected to be files when only strings being passed down to the PostEditors component. To fix this, a clear distinction was made between the file of an image and the url of an image. Now the upload of new posts work